apps/audio-api/pyproject.toml
[project]
name = "audio-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "redis",
    "beanie",
    "motor",
    "pydantic-settings",
    "shared-storage",
    "shared-messaging",
    "shared-schemas",
    "Authlib",
    "httpx",
    "Starlette",
    "sse-starlette",
    "itsdangerous",
    "google-auth-oauthlib",
    "google-api-python-client",
    "bcrypt",
    "python-docx"
]

[tool.uv.sources]
shared-storage = { workspace = true }
shared-messaging = { workspace = true }
shared-schemas = { workspace = true }